GUIFoundation: Fix Undocked Window Layout.
When restoring the layout of documents, newly created documents may not have docking information yet. We need to make sure these have a valid dockContainer so that they don't end up as floating widgets. This code tries to attach it to the settings tab, or any other document if not available.