Foundation: Prepare Expression VM For Visual Script.
Refactored the expression VM and bytecode a little bit so that it is possible to bind inputs and outputs by index instead of by name. This makes the VM setup faster and it was also easier to integrate in Visual Script.
Also had to refactor the way the bytecode is stored in memory and serialized. Now metadata and bytecode are all stored in a single memory blob and can also be loaded into an external memory blob. With that no additional allocations are needed for the bytecode when it is embedded into Visual Script.