GameEngine: Fix RenderDoc Captures In Play-The-Game Mode.
Fix Renderdoc captures in Play-the-game mode by only doing one ezGameApplication::ExecuteFrameCapture call per frame

This issue is caused by multiple windows existing at the same time (e.g. EditorView & MainWindow), so it happens that the first starts the capture and the second immediately stops the capture via the second call of ExecuteFrameCapture(), resulting in an nearly empty capture.

How to reproduce

* Open the XII Editor & Enable renderdoc
* Play-the-game (Ctrl + F5)
* Take a capture (F11).
* Open the capture in Renderdoc & observe the nearly empty capture.

With this fix, the capture contains a proper MainWindow capture.